fix courses table migration

+++ database/migrations/2019_07_22_101521_create_lessons_table.php
@@ -15,7 +15,7 @@ class CreateLessonsTable extends Migration
     {
         Schema::create('lessons', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('course_id');
+            $table->unsignedBigInteger('course_id');
             $table->unsignedSmallInteger('order');
             $table->string('title',500);
             $table->text('description');
